+++ tests/Codec/Picture/ExtraSpec.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE FlexibleContexts #-}
+
 module Codec.Picture.ExtraSpec
   ( main,
     spec,
@@ -8,6 +10,7 @@ import Codec.Picture
 import Codec.Picture.Extra
 import Control.Monad
 import Data.Function (on)
+import Foreign (Storable)
 import Test.Hspec
 
 main :: IO ()
@@ -17,6 +20,7 @@ spec :: Spec
 spec = do
   describe "scaleBilinear" scaleBilinearSpec
   describe "crop" cropSpec
+  describe "trim" trimSpec
   describe "flipHorizontally" flipHorizontallySpec
   describe "flipVertically" flipVerticallySpec
   describe "rotateLeft90Spec" rotateLeft90Spec
@@ -24,6 +28,7 @@ spec = do
   describe "rotate180" rotate180Spec
   describe "beside" besideSpec
   describe "below" belowSpec
+  describe "square" squareSpec
 
 scaleBilinearSpec :: Spec
 scaleBilinearSpec = do
@@ -78,6 +83,32 @@ cropSpec = do
         (crop 0 512 512 512)
         "data-examples/macaque-below.png"
         "data-examples/macaque.png"
+
+trimSpec :: Spec
+trimSpec = do
+  context "when we pass an image without transparency" $
+    it "does nothing" $
+      checkWithFiles
+        trim
+        "data-examples/macaque.png"
+        "data-examples/macaque.png"
+  context "when passing an image with transparency" $ do
+    it "does nothing if there are no fully transparent edges" $
+      checkWithFilesAlpha
+        trim
+        "data-examples/macaque-transparent-cropped.png"
+        "data-examples/macaque-transparent-cropped.png"
+    it "removed the transparent edges" $
+      checkWithFilesAlpha
+        trim
+        "data-examples/macaque-transparent.png"
+        "data-examples/macaque-transparent-cropped.png"
+  context "when passing a fully transparent image" $
+    it "returns an empty image" $ do
+      (Right (ImageRGBA8 original)) <- readImage "data-examples/fully-transparent.png"
+      let trimmed = trim original
+      imageWidth trimmed `shouldBe` 1
+      imageHeight trimmed `shouldBe` 1
 
 flipHorizontallySpec :: Spec
 flipHorizontallySpec =
@@ -142,6 +173,31 @@ belowSpec =
         "data-examples/macaque.png"
         "data-examples/macaque-below.png"
 
+squareSpec :: Spec
+squareSpec = do
+  context "when we pass an already square image" $ do
+    it "does nothing" $
+      checkWithFiles
+        (square $ PixelRGB8 0 0 0)
+        "data-examples/macaque.png"
+        "data-examples/macaque.png"
+    it "does nothing (alpha)" $
+      checkWithFilesAlpha
+        (square $ PixelRGBA8 0 0 0 0)
+        "data-examples/macaque-transparent.png"
+        "data-examples/macaque-transparent.png"
+  context "when we pass a non-square image" $ do
+    it "adds the required filler to wider than heigh image" $ do
+      (Right (ImageRGB8 original)) <- readImage "data-examples/macaque-beside.png"
+      let squared = square (PixelRGB8 0 0 0) original
+      imageWidth squared `shouldBe` 1024
+      imageHeight squared `shouldBe` 1024
+    it "adds the required filler to higher than wide image" $ do
+      (Right (ImageRGB8 original)) <- readImage "data-examples/macaque-below.png"
+      let squared = square (PixelRGB8 0 0 0) original
+      imageWidth squared `shouldBe` 1024
+      imageHeight squared `shouldBe` 1024
+
 -- | Run a transforming on the image loaded from a file and compare the
 -- resulting image with the contents of another file.
 checkWithFiles ::
@@ -157,6 +213,20 @@ checkWithFiles f opath fpath = do
   (Right (ImageRGB8 result)) <- readImage fpath
   f original `blindlySatisfy` sameImage result
 
+-- | `checkWithFiles` for images with an alpha channel.
+checkWithFilesAlpha ::
+  -- | Transformation to test
+  (Image PixelRGBA8 -> Image PixelRGBA8) ->
+  -- | Where to get the original image
+  FilePath ->
+  -- | Where to get the image to compare with
+  FilePath ->
+  Expectation
+checkWithFilesAlpha f opath fpath = do
+  (Right (ImageRGBA8 original)) <- readImage opath
+  (Right (ImageRGBA8 result)) <- readImage fpath
+  f original `blindlySatisfy` sameImage result
+
 -- | The same as 'shouldSatisfy', but doesn't care if its argument is an
 -- instance of 'Show' or not.
 blindlySatisfy :: a -> (a -> Bool) -> Expectation
@@ -164,7 +234,7 @@ v `blindlySatisfy` p =
   unless (p v) (expectationFailure "predicate failed")
 
 -- | The equality test for images.
-sameImage :: Image PixelRGB8 -> Image PixelRGB8 -> Bool
+sameImage :: (Eq a, Eq (PixelBaseComponent a), Storable (PixelBaseComponent a)) => Image a -> Image a -> Bool
 sameImage a b =
   ((==) `on` imageWidth) a b
     && ((==) `on` imageHeight) a b